--- ChildForm.cs
@@ -1,34 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace App_project
-{
-    class ChildForm
-    {
-       
-        private static void OpenChildForm(Form childForm, Form panel)
-        {
-            Form activeForm = null;
-            if (activeForm != null)
-            {
-                activeForm.Close();
-            }
-            activeForm = childForm;
-
-            childForm.TopLevel = false;
-            childForm.FormBorderStyle = FormBorderStyle.None;
-            childForm.Dock = DockStyle.Fill;
-
-            panel.Controls.Add(childForm);
-            panel.Tag = childForm;
-
-            childForm.BringToFront();
-            childForm.Show();
-        }
-
-    }
-}
