+++ Глобальные переменные/IDPetCard_key.cs
