+++ Глобальные переменные/IDUser_key.cs
