--- Сортировка/ISort.cs
@@ -1,11 +1,0 @@
-﻿using System;
-namespace App_project
-{
-    /// 
-    /// Интерфейс для сортировки
-    /// 
-    public interface ISort
-    {
-         void ToSort();
-    }
-}
